comminting java on 22/3

+++ TimeSheetManagement/account_service/src/main/java/com/feuji/accountservice/dto/UpdateAccountDto.java
@@ -20,7 +20,7 @@ public class UpdateAccountDto {
 	private String businessDevelopmentManagerName;
 	private Integer parentAccountId;
 	private String parentAccountName;
-	private Long accountBuId;
+	private Integer accountBuId;
 	private String accountBuName;
 	private Timestamp plannedStartDate;
 	private Timestamp  plannedEndDate;
@@ -31,7 +31,7 @@ public class UpdateAccountDto {
 	private String state;
 	private String zipcode;
 	private String country;
-	private Long statusId;
+	private Integer statusId;
 	private String statusValue;
 	private String comments;
 	private String uuId;
@@ -63,10 +63,10 @@ public class UpdateAccountDto {
 
 	public UpdateAccountDto(Integer accountId, String accountName, Integer ownerId, String ownerName,
 			Integer relationshipManagerId, String relationshipManageName, Integer businessDevelopmentManagerId,
-			String businessDevelopmentManagerName, Integer parentAccountId, String parentAccountName, Long accountBuId,
+			String businessDevelopmentManagerName, Integer parentAccountId, String parentAccountName, Integer accountBuId,
 			String accountBuName, Timestamp plannedStartDate, Timestamp plannedEndDate, Timestamp actualStartDate,
 			Timestamp actualEndDate, String address, String city, String state, String zipcode, String country,
-			Long statusId, String statusValue, String comments, String uuId) {
+			Integer statusId, String statusValue, String comments, String uuId) {
 		super();
 		this.accountId = accountId;
 		this.accountName = accountName;
